android/requirements.txt
kivy==2.3.1
kivymd==1.1.1
kivy_garden.graph
python-dateutil
pillow
buildozer
